# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API TechUniverse",
        "description": "Documentacion de la API TechUniverse",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://ivan123.alwaysdata.net",
            "description": "Servidor principal"
        }
    ],
    "paths": {
...
        "/api/productos": {
            "get": {
                "tags": [
                    "Productos"
                ],
                "summary": "Listar productos con paginación",
                "description": "Retorna una lista paginada de productos, incluyendo su categoría y especificaciones.",
                "operationId": "70a301af8ea04ce2f2596a58758632cd",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Número de página a obtener",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista paginada de productos obtenida correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Producto"
                                            }
                                        },
                                        "meta": {
                                            "properties": {
                                                "current_page": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "last_page": {
                                                    "type": "integer",
                                                    "example": 5
                                                },
                                                "per_page": {
                                                    "type": "integer",
                                                    "example": 12
                                                },
                                                "total": {
                                                    "type": "integer",
                                                    "example": 60
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Productos"
                ],
                "summary": "Crear un nuevo producto",
                "description": "Crea un producto nuevo con imagen subida a R2",
                "operationId": "4ff8851cb6bfe92aa226cdc804917ff0",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "nombre",
                                    "precio",
                                    "categoria_id",
                                    "foto"
                                ],
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "Laptop Gaming"
                                    },
                                    "precio": {
                                        "type": "integer",
                                        "minimum": 1,
                                        "example": 1500
                                    },
                                    "descripcion": {
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "Descripción opcional",
                                        "nullable": true
                                    },
                                    "categoria_id": {
                                        "type": "integer",
                                        "minimum": 1,
                                        "example": 2
                                    },
                                    "foto": {
                                        "description": "Imagen del producto (máx. 4MB)",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Producto creado con éxito",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "mensaje": {
                                            "type": "string",
                                            "example": "Producto creado con éxito"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Producto"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The nombre field is required."
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/productos/{id}": {
            "get": {
                "tags": [
                    "Productos"
                ],
                "summary": "Obtener un producto por ID",
                "description": "Retorna un producto específico con su categoría y especificaciones",
                "operationId": "1884f1fc784a80f4976ac9c8d4ce887b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del producto",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Producto encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Producto"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Producto no encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "No encontrado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Productos"
                ],
                "summary": "Actualizar un producto",
                "description": "Actualiza los datos de un producto. Se usa POST con _method=PUT por compatibilidad con multipart/form-data",
                "operationId": "308ad2424a4756367ccf91c0bc1dcafc",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del producto a actualizar",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "_method": {
                                        "description": "Spoofing de método HTTP",
                                        "type": "string",
                                        "default": "PUT"
                                    },
                                    "nombre": {
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "Laptop Pro",
                                        "nullable": true
                                    },
                                    "precio": {
                                        "type": "integer",
                                        "minimum": 1,
                                        "example": 2000,
                                        "nullable": true
                                    },
                                    "descripcion": {
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "Nueva descripción",
                                        "nullable": true
                                    },
                                    "categoria_id": {
                                        "type": "integer",
                                        "minimum": 1,
                                        "example": 3,
                                        "nullable": true
                                    },
                                    "stock": {
                                        "type": "integer",
                                        "minimum": 0,
                                        "example": 10,
                                        "nullable": true
                                    },
                                    "foto": {
                                        "description": "Nueva imagen (máx. 4MB)",
                                        "type": "string",
                                        "format": "binary",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Producto actualizado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "mensaje": {
                                            "type": "string",
                                            "example": "Actualizado correctamente"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Producto"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Producto no encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "No encontrado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The precio must be an integer."
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Productos"
                ],
                "summary": "Eliminar un producto",
                "description": "Elimina un producto y su imagen del bucket R2",
                "operationId": "ea733033fbff1681e04f8c4d0e11e0bc",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del producto a eliminar",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Producto eliminado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "mensaje": {
                                            "type": "string",
                                            "example": "Eliminado correctamente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Producto no encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "No encontrado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Producto": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "nombre": {
                        "type": "string",
                        "example": "Laptop Gaming"
                    },
                    "precio": {
                        "type": "integer",
                        "example": 1500
                    },
                    "descripcion": {
                        "type": "string",
                        "example": "Descripción del producto",
                        "nullable": true
                    },
                    "foto": {
                        "type": "string",
                        "example": "https://cdn.example.com/productos/laptop.jpg"
                    },
                    "categoria": {
                        "properties": {
                            "id": {
                                "type": "integer",
                                "example": 2
                            },
                            "nombre": {
                                "type": "string",
                                "example": "Electrónica"
                            }
                        },
                        "type": "object"
                    },
                    "especificaciones": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "id": {
                                    "type": "integer",
                                    "example": 1
                                },
                                "nombre": {
                                    "type": "string",
                                    "example": "RAM"
                                },
                                "valor": {
                                    "type": "string",
                                    "example": "16GB"
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}